Reorganized folders. Set up alias for asset import
