+++ client/src/features/Cricket/leagues/my-team/subroutes/Results/components/statCard.tsx
