+++ client/tsconfig.json
@@ -3,5 +3,12 @@
     "references": [
         { "path": "./tsconfig.app.json" },
         { "path": "./tsconfig.node.json" }
-    ]
+    ],
+    "compilerOptions": {
+    "baseUrl": "src", 
+    "paths": {
+      "@assets/*": ["assets/*"]
+    }
+  }
+
 }
