+++ client/vite.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from "vite";
 import topLevelAwait from "vite-plugin-top-level-await";
 import wasm from "vite-plugin-wasm";
 import fs from "fs";
+import path from 'path';
 
 export default defineConfig({
        server: {
@@ -13,4 +14,9 @@ export default defineConfig({
          },
       },
     plugins: [react(), wasm(), topLevelAwait()],
+    resolve: {
+      alias: {
+        "@assets": path.resolve(__dirname, "src/assets"),
+      },
+    },
 });
